chat-service/pyproject.toml
[project]
name = "chat-service"
version = "0.1.0"
description = "chat microservice"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-socketio==5.11.4",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.12",
    "motor==3.6.0",
    "redis==5.2.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "python-dotenv==1.0.1",
    "aioredis==2.0.1",
    "pymongo==4.10.1",
    "websockets==13.1",
]
